+++ ocean_provider/routes/consume.py
@@ -14,9 +14,11 @@ from ocean_provider.utils.basics import (
     LocalFileAdapter,
     get_provider_wallet,
     get_web3,
+    get_metadata_url,
     validate_timestamp,
 )
 from ocean_provider.utils.compute_environments import check_environment_exists
+from ocean_provider.utils.datatoken import validate_order
 from ocean_provider.utils.error_responses import error_response
 from ocean_provider.utils.provider_fees import get_provider_fees, get_c2d_environments
 from ocean_provider.utils.services import ServiceType
@@ -25,10 +27,8 @@ from ocean_provider.utils.util import (
     build_download_response,
     check_asset_consumable,
     get_download_url,
-    get_metadata_url,
     get_request_data,
     get_service_files_list,
-    validate_order,
     validate_url_object,
 )
 from ocean_provider.validation.provider_requests import (
